moved supporting code to libs, updated readme, added required gems
